+++ packages/graphql-server/src/index.ts
@@ -17,7 +17,7 @@ import type {
   GraphQLFormattedError,
 } from 'graphql'
 
-import type { Context, Env, Input } from 'hono'
+import type { Context, Env, Input, MiddlewareHandler } from 'hono'
 import { parseBody } from './parse-body'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -33,18 +33,18 @@ type Options<E extends Env = any, P extends string = any, I extends Input = {}> 
   rootResolver?: RootResolver<E, P, I>
   pretty?: boolean
   validationRules?: ReadonlyArray<ValidationRule>
-  // graphiql?: boolean
+  graphiql?: boolean
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const graphqlServer = <E extends Env = any, P extends string = any, I extends Input = {}>(
   // eslint-enable-next-line @typescript-eslint/no-explicit-any
   options: Options<E, P, I>
-) => {
+): MiddlewareHandler => {
   const schema = options.schema
   const pretty = options.pretty ?? false
   const validationRules = options.validationRules ?? []
-  // const showGraphiQL = options.graphiql ?? false
+  const showGraphiQL = options.graphiql ?? false
 
   return async (c: Context<E, P, I>) => {
     // GraphQL HTTP only supports GET and POST methods.
@@ -68,6 +68,9 @@ export const graphqlServer = <E extends Env = any, P extends string = any, I ext
     const { query, variables, operationName } = params
 
     if (query == null) {
+      if (showGraphiQL && c.req.method === 'GET') {
+        return respondWithGraphiQL(c)
+      }
       return c.json(errorMessages(['Must provide query string.']), 400)
     }
 
@@ -107,13 +110,6 @@ export const graphqlServer = <E extends Env = any, P extends string = any, I ext
       // Determine if this GET request will perform a non-query.
       const operationAST = getOperationAST(documentAST, operationName)
       if (operationAST && operationAST.operation !== 'query') {
-        /*
-        Now , does not support GraphiQL
-        if (showGraphiQL) {
-          //return respondWithGraphiQL(response, graphiqlOptions, params)
-        }
-        */
-
         // Otherwise, report a 405: Method Not Allowed error.
         return c.json(
           errorMessages([
@@ -155,12 +151,6 @@ export const graphqlServer = <E extends Env = any, P extends string = any, I ext
         return c.json(errorMessages([result.errors.toString()], result.errors), 500)
       }
     }
-
-    /*
-    Now, does not support GraphiQL
-    if (showGraphiQL) {
-    }
-    */
 
     if (pretty) {
       const payload = JSON.stringify(result, null, pretty ? 2 : 0)
@@ -242,4 +232,87 @@ export const errorMessages = (
   }
 }
 
-// export const graphiQLResponse = () => {}
+export const respondWithGraphiQL = (c: Context) => {
+  // https://github.com/graphql/graphiql/blob/85edb9e0505db8ff963c9ad4674bc8fa2e02a35a/examples/graphiql-cdn/index.html
+  return c.html(`<!--
+*  Copyright (c) 2021 GraphQL Contributors
+*  All rights reserved.
+*
+*  This source code is licensed under the license found in the
+*  LICENSE file in the root directory of this source tree.
+-->
+<!doctype html>
+<html lang="en">
+  <head>
+    <title>GraphiQL</title>
+    <style>
+      body {
+        height: 100%;
+        margin: 0;
+        width: 100%;
+        overflow: hidden;
+      }
+
+      #graphiql {
+        height: 100vh;
+      }
+    </style>
+    <!--
+      This GraphiQL example depends on Promise and fetch, which are available in
+      modern browsers, but can be "polyfilled" for older browsers.
+      GraphiQL itself depends on React DOM.
+      If you do not want to rely on a CDN, you can host these files locally or
+      include them directly in your favored resource bundler.
+    -->
+    <script
+      crossorigin
+      src="https://unpkg.com/react@18/umd/react.development.js"
+    ></script>
+    <script
+      crossorigin
+      src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"
+    ></script>
+    <!--
+      These two files can be found in the npm module, however you may wish to
+      copy them directly into your environment, or perhaps include them in your
+      favored resource bundler.
+    -->
+    <script
+      src="https://unpkg.com/graphiql/graphiql.min.js"
+      type="application/javascript"
+    ></script>
+    <link rel="stylesheet" href="https://unpkg.com/graphiql/graphiql.min.css" />
+    <!--
+      These are imports for the GraphIQL Explorer plugin.
+    -->
+    <script
+      src="https://unpkg.com/@graphiql/plugin-explorer/dist/index.umd.js"
+      crossorigin
+    ></script>
+
+    <link
+      rel="stylesheet"
+      href="https://unpkg.com/@graphiql/plugin-explorer/dist/style.css"
+    />
+  </head>
+
+  <body>
+    <div id="graphiql">Loading...</div>
+    <script>
+      const root = ReactDOM.createRoot(document.getElementById('graphiql'));
+      const fetcher = GraphiQL.createFetcher({
+        url: '${c.req.path}',
+      });
+      const explorerPlugin = GraphiQLPluginExplorer.explorerPlugin();
+      root.render(
+        React.createElement(GraphiQL, {
+          fetcher,
+          defaultEditorToolsVisibility: true,
+          plugins: [explorerPlugin],
+        }),
+      );
+    </script>
+  </body>
+</html>
+`)
+}