+++ packages/graphql-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hono/graphql-server
 
+## 0.5.0
+
+### Minor Changes
+
+- [#670](https://github.com/honojs/middleware/pull/670) [`e3eafbc117103ef408f8b242784b53c9ffdb326e`](https://github.com/honojs/middleware/commit/e3eafbc117103ef408f8b242784b53c9ffdb326e) Thanks [@3w36zj6](https://github.com/3w36zj6)! - feat: add GraphiQL support
+
+## 0.4.4
+
+### Patch Changes
+
+- [#667](https://github.com/honojs/middleware/pull/667) [`41ab4b029d225901b7369d88fb29b64f6c217dda`](https://github.com/honojs/middleware/commit/41ab4b029d225901b7369d88fb29b64f6c217dda) Thanks [@3w36zj6](https://github.com/3w36zj6)! - fix: add type to Middleware Handler
+
 ## 0.4.3
 
 ### Patch Changes